requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
backoff==1.11.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
dnspython==2.7.0
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.8.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.1
propcache==0.4.1
pymongo==4.16.0
python-dotenv==1.2.1
redis==7.0.1
requests==2.32.5
spotipy==2.25.2
SQLAlchemy==2.0.46
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
WTForms==3.2.1
yarl==1.22.0
zipp==3.23.0
